move helpers into util and actions to top of module
