+++ test/test_db_control.py
@@ -8,29 +8,6 @@ import forest.db.control
 from forest import db, redux, rx
 
 
-@pytest.mark.parametrize("given,expect", [
-    pytest.param(
-        dt.datetime(2020, 1, 1),
-        dt.datetime(2021, 1, 1),
-        id="datetime.datetime"),
-    pytest.param(
-        np.datetime64("2020-01-01 00:00:00", "s"),
-        np.datetime64("2021-01-01 00:00:00", "s"),
-        id="datetime64[s]"),
-    pytest.param(
-        "2020-01-01 00:00:00",
-        "2021-01-01 00:00:00",
-        id="str"),
-    pytest.param(
-        cftime.DatetimeGregorian(2020, 1, 1),
-        cftime.DatetimeGregorian(2021, 1, 1),
-        id="cftime.DatetimeGregorian"),
-])
-def test_replace(given, expect):
-    result = forest.db.control.replace(given, year=2021)
-    assert result == expect
-
-
 def test_reducer_immutable_state():
     """Ensure copy.deepcopy is used to create a new state"""
     previous_state = {"key": ["value"]}
