+++ test/test_util.py
@@ -54,3 +54,26 @@ class Test_to_datetime(unittest.TestCase):
 ])
 def test_parse_date(regex, fmt, path, expect):
     assert util.parse_date(regex, fmt, path) == expect
+
+
+@pytest.mark.parametrize("given,expect", [
+    pytest.param(
+        dt.datetime(2020, 1, 1),
+        dt.datetime(2021, 1, 1),
+        id="datetime.datetime"),
+    pytest.param(
+        np.datetime64("2020-01-01 00:00:00", "s"),
+        np.datetime64("2021-01-01 00:00:00", "s"),
+        id="datetime64[s]"),
+    pytest.param(
+        "2020-01-01 00:00:00",
+        "2021-01-01 00:00:00",
+        id="str"),
+    pytest.param(
+        cftime.DatetimeGregorian(2020, 1, 1),
+        cftime.DatetimeGregorian(2021, 1, 1),
+        id="cftime.DatetimeGregorian"),
+])
+def test_replace(given, expect):
+    result = util.replace(given, year=2021)
+    assert result == expect
